requirements.txt
mmh3
pyyaml